SCANMEIFUCAN_Backend/requirements.txt
Flask
flask-cors
Flask-SQLAlchemy
psycopg2-binary
pytesseract==0.3.10
Pillow==10.2.0
numpy